moved declaration of "class Implementation" to models.h and images.h
