+++ src/osg/BuildOptions.h
@@ -1,23 +1,13 @@
 #pragma once
 
 #include <vsg/all.h>
+#include <vsgXchange/models.h>
 
 #include "GeometryUtils.h"
 #include "ShaderUtils.h"
 
 namespace osg2vsg
 {
-    struct PipelineCache : public vsg::Inherit<vsg::Object, PipelineCache>
-    {
-        using Key = std::tuple<uint32_t, uint32_t, std::string, std::string>;
-        using PipelineMap = std::map<Key, vsg::ref_ptr<vsg::BindGraphicsPipeline>>;
-
-        std::mutex mutex;
-        PipelineMap pipelineMap;
-
-        vsg::ref_ptr<vsg::BindGraphicsPipeline> getOrCreateBindGraphicsPipeline(uint32_t shaderModeMask, uint32_t geometryMask, const std::string& vertShaderPath = "", const std::string& fragShaderPath = "");
-    };
-
     struct BuildOptions : public vsg::Inherit<vsg::Object, BuildOptions>
     {
         vsg::ref_ptr<const vsg::Options> options;
